backend-api/requirements.txt
# Backend API dependencies
fastapi==ç
uvicorn==0.24.0
SQLAlchemy==2.0.22
bcrypt==4.1.2
passlib==1.7.4
PyJWT==2.8.0
python-dotenv==1.2.1
python-multipart==0.0.6
email-validator==2.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
